Move forming of sidebar into a hook.
